fixed start app, restructured project
