+++ tests/api/test_add_reserve.py
@@ -3,7 +3,6 @@ from datetime import datetime, timedelta
 import pytest
 
 
-
 @pytest.mark.asyncio
 async def test_create_reservation_and_duplicate(async_client, prepare_table: int):
     reservation_time = (datetime.now() + timedelta(hours=1)).replace(microsecond=0)
@@ -12,7 +11,7 @@ async def test_create_reservation_and_duplicate(async_client, prepare_table: int
         "customer_name": "test_name",
         "table_id": prepare_table,
         "reservation_time": reservation_time.isoformat(),
-        "duration_minutes": 60
+        "duration_minutes": 60,
     }
 
     response_1 = await async_client.post("/reservations/", json=new_reservation_data)
@@ -25,7 +24,6 @@ async def test_create_reservation_and_duplicate(async_client, prepare_table: int
     assert response_2.json() == {"detail": "Столик на это время уже забронирован"}
 
 
-
 @pytest.mark.asyncio
 async def test_create_reservation_in_the_past(async_client, prepare_table: int):
     reservation_time = (datetime.now() - timedelta(hours=2)).replace(microsecond=0)
@@ -34,7 +32,7 @@ async def test_create_reservation_in_the_past(async_client, prepare_table: int):
         "customer_name": "past_client",
         "table_id": prepare_table,
         "reservation_time": reservation_time.isoformat(),
-        "duration_minutes": 30
+        "duration_minutes": 30,
     }
 
     response = await async_client.post("/reservations/", json=new_reservation_data)
