+++ tests/api/test_delete_table.py
@@ -1,4 +1,3 @@
-
 import pytest
 
 from db.connector import DatabaseConnector
@@ -6,7 +5,9 @@ from db.models import Table
 
 
 @pytest.mark.asyncio
-async def test_delete_table_204(async_client, dine_easy_db: DatabaseConnector, prepare_table):
+async def test_delete_table_204(
+    async_client, dine_easy_db: DatabaseConnector, prepare_table
+):
     table_id = prepare_table
     response = await async_client.delete(f"/tables/{table_id}")
     assert response.status_code == 204, response.text
@@ -19,6 +20,6 @@ async def test_delete_table_204(async_client, dine_easy_db: DatabaseConnector, p
 
 @pytest.mark.asyncio
 async def test_delete_table_404(async_client):
-    response = await async_client.delete(f"/tables/10002")
+    response = await async_client.delete("/tables/10002")
     assert response.status_code == 404, response.text
     assert response.json() == {"detail": "Столик не найден"}
