+++ tests/api/test_get_reserve_list.py
@@ -4,7 +4,9 @@ from db.models import Reservation
 
 
 @pytest.mark.asyncio
-async def test_get_reservations(async_client, prepare_tables, prepare_reservations, reservations: list[Reservation]):
+async def test_get_reservations(
+    async_client, prepare_tables, prepare_reservations, reservations: list[Reservation]
+):
     response = await async_client.get("/reservations/")
 
     assert response.status_code == 200, response.text
@@ -15,7 +17,7 @@ async def test_get_reservations(async_client, prepare_tables, prepare_reservatio
             "customer_name": reservation.customer_name,
             "table_id": reservation.table_id,
             "reservation_time": reservation.reservation_time.isoformat(),
-            "duration_minutes": reservation.duration_minutes
+            "duration_minutes": reservation.duration_minutes,
         }
         for reservation in reservations
     ]
